# swagger.json
{
  "info": {
    "description": "Test API with autogenerated swagger doc",
    "swagger": "2.0",
    "title": "Node Express API example",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
    "/auth": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "patch user by id.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "email example",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "example": "test123",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "patch user by id.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "name example",
                  "required": true
                },
                "email": {
                  "type": "string",
                  "example": "email example",
                  "required": true
                },
                "address": {
                  "type": "string",
                  "example": "address example",
                  "required": true
                },
                "ktp": {
                  "type": "string",
                  "example": "ktp example",
                  "required": true
                },
                "npwp": {
                  "type": "string",
                  "example": "npwp example",
                  "required": true
                },
                "passport": {
                  "type": "string",
                  "example": "passport example",
                  "required": true
                },
                "password": {
                  "type": "string",
                  "example": "test123",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "patch user by id.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "string"
          },
          {
            "in": "query",
            "name": "rows",
            "type": "string"
          },
          {
            "in": "query",
            "name": "orderby",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sort",
            "type": "string"
          },
          {
            "in": "query",
            "name": "name",
            "type": "string"
          },
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "address",
            "type": "string"
          },
          {
            "in": "query",
            "name": "ktp",
            "type": "string"
          },
          {
            "in": "query",
            "name": "npwp",
            "type": "string"
          },
          {
            "in": "query",
            "name": "passport",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "patch user by id.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "name example"
                },
                "email": {
                  "type": "string",
                  "example": "email example"
                },
                "address": {
                  "type": "string",
                  "example": "address example"
                },
                "ktp": {
                  "type": "string",
                  "example": "ktp example"
                },
                "npwp": {
                  "type": "string",
                  "example": "npwp example"
                },
                "passport": {
                  "type": "string",
                  "example": "passport example"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    },
    "/users/:mailorname": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "get user by mail or name.",
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    },
    "/users/versions/:id": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "get user versions by id.",
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    },
    "/users/:id": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "patch user by id.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "name example"
                },
                "email": {
                  "type": "string",
                  "example": "email example"
                },
                "address": {
                  "type": "string",
                  "example": "address example"
                },
                "ktp": {
                  "type": "string",
                  "example": "ktp example"
                },
                "npwp": {
                  "type": "string",
                  "example": "npwp example"
                },
                "passport": {
                  "type": "string",
                  "example": "passport example"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "soft deletes user by id.",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK."
          },
          "400": {
            "description": "Bad Request."
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}